Use long instead of int for malloc size

+++ latte/templates/aligned_malloc.c
@@ -1,4 +1,4 @@
 extern "C"
-void * aligned_malloc(int size) {
+void * aligned_malloc(long size) {
     return (void *) _mm_malloc(size, 64);
 }
